requirements.txt
PyYAML==6.0
attrs==21.2.0
